Rename Infinispan FAT package
