+++ dev/io.openliberty.jcache.internal_fat.infinispan/fat/src/io/openliberty/jcache/internal/infinispan/fat/docker/InfinispanContainer.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package io.openliberty.jcache.internal.fat.docker;
+package io.openliberty.jcache.internal.infinispan.fat.docker;
 
 import static org.hamcrest.CoreMatchers.anyOf;
 import static org.hamcrest.CoreMatchers.is;
